Upgrade `feePayer` to a signer when using `addSignersToTransactionMessage` (#151)
This PR upgrades the `feePayer` attribute of a transaction message if and only if:
- The current `feePayer` attribute is set.
- The current `feePayer` attribute is not already a signer (i.e. address only).
- The current `feePayer` address matches one of the provided signers.

Note that the `addSignersToTransactionMessage` does not narrow the returned transaction message with a `ITransactionMessageWithFeePayerSigner` because this may or may not happen based on the provided signers. If desired, we could iterate over the explicit `TransactionSigner<TAddress>` array to figure out at compile time if we need to return a `ITransactionMessageWithFeePayerSigner` type. This felt like an overkill and over-complication of the types to me but happy to disagree and commit.

Fixes #82