Always compare the current PR's bundle size to main (#218)
Regardless of what the mergebase of the current PR is, this will cause Bundlemon to _always_ compare the size from the perspective of the last known `main` size.

This prevents stacked PRs from trying (and failing) to find the size of the PR branch directly underneath them.

Fixes #70

# Test Plan

Works on #72.