Add docs website (#305)
This PR moves the documentation website from a separate repo to a `docs` folder inside the main repo. This makes it possible to update documentation and code within the same PR.

Note that both the design and content of the documentation website is still a work in progress.

P.S.: We need to run `pnpm install --ignore-workspace` inside the `docs` folder locally in order to avoid mixing the dependencies with the rest of the library. Let me know if you'd rather we tackle this differently.