A package for creating nominal TypeScript types (#472)
#### Problem

Two versions of any given package may introduce two distinct instances of our branded types into a project. That is to say you might get the `Address` type from v2.1.0 and v3.0.0 in the same project.

When this happens, you will end up with values that are runtime identical, but whose types mismatch. This is because the way we brand the types involves `unique symbol` which is locally unique to the particular instance of the package being imported.

#### Summary of Changes

Create a package with utilities that will brand types with shared names across the entire project, instead of unique symbols that are module-local.

Addresses #459