Modify all branded types such that they will match with any structurally similar type from any package version, going forward (#473)
#### Problem

Two versions of any given package may introduce two distinct instances of our branded types into a project. That is to say you might get the `Address` type from v2.0.0 and v2.1.0 in the same project.

#### Summary of Changes

Here we use the type utilities in the previous stacked PR to reimagine these types as structurally identical, rather than having an identity that's tied to one particular instance of a module.

This will make it so that `Address` and other branded types like it are compatible from version v2.1.1 and onward.

Fixes #459