Add transaction size helpers (#425)
This PR adds various types and functions to help get the size of a transaction (or its message) and ensure its size is valid (under the transaction size limit).

Additionally, it provides two new types — `TransactionUnderSizeLimit` and `TransactionMessageUnderSizeLimit` — to provide additional type safety before sending a transaction.

In this PR, `TransactionUnderSizeLimit` and `TransactionMessageUnderSizeLimit` are only used by their respective type guards but this stack of PR aims to use these types to offer more type safety in other parts of the code. For instance: adding an instruction to or changing the lifetime of a transaction message should remove the `TransactionMessageUnderSizeLimit` type if any; sending a transaction should require the `TransactionUnderSizeLimit`; compiling a `TransactionMessageUnderSizeLimit` into a transaction should return a `TransactionUnderSizeLimit`; etc.

The next few PRs don't do any of that just yet but prepare the field for such changes. Mainly, it aims to bring more consistency amongst the type flags of transactions and transaction messages so adding yet another one does not affect the developer experience.