Add transaction message Exclude type helpers (#430)
This PR introduce a `TransactionMessageWithLifetime` type that describes a transaction message with any lifetime set.

It also introduce `ExcludeTransactionMessageX<TTransactionMessage>` types for most `TransactionMessageWithX` types. This means we can simplify the return type of most of our setters by `ExcludeTransactionMessageX<T> & TransactionMessageWithX` to ensure the old state is not kept.

For instance, see how this PR makes the `setTransactionMessageLifetimeUsingBlockhash` function use a single function signature. This will be useful when we'll need to chain more type exclusions to setters such as `ExcludeTransactionMessageUnderSizeLimit`.