Make rpc non-thenable (#508)
#### Problem
RPC objects created by Solana Kit use a proxy that converts all property access into methods. When JavaScript's Promise machinery checks for a `.then` property (which happens when you `async await` an RPC object), the proxy intercepts this and returns a method `{ send: [AsyncFunction: send] }` instead of `undefined` or the proper Promise interface, making the RPC object appear thenable when it isn't. This causes JavaScript/TypeScript to silently terminate the program when you try to `await` the RPC object as if it were a Promise.

#### Summary of Changes
- Modified `makeProxy` function to specifically intercept `then` property access and return `undefined`
- Prevents the proxy from treating `.then` like any other RPC method
- Added test case to verify RPC objects are not thenable