Add `TransactionWithLifetime` requirement to signing functions (#574)
This PR adds the `TransactionWithLifetime` requirement when signing transactions — including the `TransactionPartialSigner`.

This is because, when signing a transaction, we must know its final lifetime constraint as any further update will invalidate signatures.

Note that we did not update the `TransactionModifyingSigner` nor the `TransactionSendingSigner` as they have the ability to transform transactions before singing them, which includes adding a lifetime.