Extract lifetime from `CompiledTransactionMessage` (#594)
This PR removes the `lifetimeToken` attribute from `CompiledTransactionMessage` and adds it to a new type `CompiledTransactionMessageWithLifetime` that can be composed with the former.

This enables `CompiledTransactionMessages` to be encoded without the need to specify a mock lifetime token and for decoded compiled messages to be typed as having a lifetime.