Add `createBaseTransactionPlanner` function (#648)
This PR provides a base implementation for the `TransactionPlanner` type that tries to pack transactions in the most optimal way possible.

This function accepts a `createTransactionMessage` callback that dictates how each new transaction message should be created before filling them with instructions. Here's an example:

```ts
const planner = createBaseTransactionPlanner({
  createTransactionMessage: () =>
    pipe(
      createTransactionMessage({ version: 0 }),
      (message) => setTransactionMessageFeePayerSigner(payer, message),
      (message) => setTransactionMessageComputeUnitPrice(cuPrice, message),
      // ...
    ),
});

const transactionPlan = await planner(instructionPlan);
```

An optional `onTransactionMessageUpdated` callback can also be provided and will be executed any time something changes in a transaction message being planned (for instance, when new instructions are being added to the message). This allows users to hook logic into the planning process.

Note, the function is called `createBaseTransactionPlanner` because I was initially planning on having a high-level `createDefaultTransactionPlanner` that is all battery included but I'm thinking this interface is already good enough to be used as a high-level. So **maybe I should rename this to `createTransactionPlanner`? Wdyt?**