[v3] Make `FullySignedTransaction` composable (#462)
#### Problem

The `FullySignedTransaction` type extends the `Transaction` interface by simply adding a `unique symbol` flagging it as fully signed. This would be great if this was the only flag we wanted to add to `Transaction` but this is no longer the case as we are planning to add `TransactionWithinSizeLimit` as well.

#### Summary of Changes

This PR plans for the addition of `TransactionWithinSizeLimit` by making the `FullySignedTransaction` more composable. Instead of inheriting from the `Transaction` type, it becomes this standalone flag that can be added to `Transaction` types to narrow them.