Use canary version in documentation website (#736)
This PR updates the Solana Kit packages imported in our documentation to the latest canary version and adjusts the content and code examples accordingly.

This is important so our documentation as access to Instruction Plan types in our order to document this concept in a subsequent PR.

This PR also makes https://github.com/anza-xyz/kit/pull/585 redundant.