Avoid type cast in getting started guide (#739)
This PR re-writes the "Set the compute limit dynamically" section of the getting started guide in order to avoid having a forced type-cast in the documentation — i.e. `as unknown as T`.

It does this by implementing a factory function separately first and using its return type to set the new attribute on the `Client` type. That way, the function type is precise and automatically set by TS.