Use same sidebar on API and docs layouts (#744)
This PR fixes the API layout of the documentation such that it matches the ones used by the "Docs" layout — that is, the part of the documentation that contains guides.

In other words, it ensures the sidebar and navbar are the same regardless of whether we are in the "Documentation" area or the "API References" area.