Generate API index page (#746)
This was a fun one.

This PR generates content for the welcome page of the "API References" area of the documentation website.

I went a bit fancy and decided to generate and parse the TypeDocs JSON files for every package so that we could auto-generate links to the existing pages.

![CleanShot 2025-08-14 at 16.45.54@2x.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/kTNWsdvmzyPPLaKihnPF/2923ca1c-934a-4020-b253-3e252d4943c0.png)

I also made sure that, if an item was exported from package `A` but also imported from package `B`, then:

- it only shows as an item on package `B`.
- package `B` shows in the "packages" list of package `A`.

Notice how the `@solana/kit` package shows the packages it re-exports but also the ones that only it exports.

![CleanShot 2025-08-14 at 16.55.14@2x.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/kTNWsdvmzyPPLaKihnPF/4070b4af-bc9e-493e-b468-27651b1762f8.png)