Server render all of the homepage examples (#850)
#### Problem

We're importing `.source/index.ts` in a client component, which means that we're getting the entire MDX runtime on the client (ie. on the homepage). This is bad for two reasons.

1. Increased bundle size on the homepage
2. We now depend on node browser-side polyfills for stuff in the MDX runtime (like `node:fs`). This also consequently makes it impossible to upgrade next.js.

Unfortunately, just straight up server rendering all of the examples and passing them down as props to the `<CodeContent />` component doesn't seem to work either, presumably because Twoslash components are themselves client components, and they’re missing type information or something in the browser, making half of the code disappear when rendered.

![image.png](https://app.graphite.dev/user-attachments/assets/ca1f3425-23c6-41d5-a103-0644aae12f34.png)

#### Summary of Changes

- Remove twoslash markup from the homepage examples
- Server render all of the homepage examples and pass them down as props