Add a function to extract the lifetime from a CompiledTransactionMessage (#918)
#### Problem

When writing a signer, it is possible that you receive a Transaction from for example an external wallet, and need to determine its lifetime constraint.

#### Summary of Changes

This PR adds a new function `getTransactionLifetimeConstraintFromCompiledTransactionMessage` which enables a best-effort extraction of the lifetime constraint from a `CompiledTransactionMessage`, ie without full information about the transaction.

The check is less robust than the equivalent for a `TransactionMessage`, because the accounts passed to an advance nonce instruction may be in lookup tables - and therefore we can't verify their address or role. However, we are able to check the program address, program data and number of accounts passed in. A transaction that incorrectly passes this check is one that passes incorrect accounts to the advance nonce instruction. The impact of attempting to confirm it with an incorrect lifetime should be minimal (it will fail to land regardless).

There is one error case I've added - if the first account passed to the instruction is not a static address, then we can't use its address in the returned `TransactionDurableNonceLifetime` and we throw. I think this is an edge case and not worth the cost of resolving lookup tables.

This function will be used to enable `useWalletAccountTransactionSigner` to return a lifetime in cases where the wallet modifies the lifetime token of the input transaction. I think it will also be useful to anybody else writing similar signers.