Add functions to narrow a TransactionWithLifetime to a specific lifetime (#925)
#### Summary of Changes

This PR adds functions to narrow from a `Transaction | (Transaction & TransactionWithLifetime)` to `TransactionWithBlockhashLifetime` or `TransactionWithDurableNonceLifetime`.

These functions will be useful when we update the modifying signer API to return `TransactionWithLifetime` instead of mirroring the input transaction message type. It will then be necessary to use one of these functions to narrow the type if you wish to use Kit's confirmation logic, which depends on a specific `lifetimeConstraint` being attached to transactions.