Add a function to create a decoder that checks the size of the input bytes (#944)
#### Problem

When an app has access to many codecs, for example when using generated clients, it is easy to write code that accidentally uses the wrong type of `Decoder` for a given byte array. This is difficult to debug as you will often be able to use the wrong decoder without any errors, but will be getting invalid data.

#### Summary of Changes

This PR adds a new helper function `createDecoderThatUsesExactByteArray` that can wrap any decoder. The transformed decoder throws if the new offset after decoding a byte array, is not at the end of the input bytes.

This means that if you pass a byte array that is longer or shorter than the wrapped decoder expects, the decode will fail.

As noted in the issue this is not a strong guarantee that you're using the correct decoder, many decoders are the same size or variable size. But it should catch a lot of difficult to debug cases in practice, such as decoding account data using the wrong decoder.

Note that to enable this, the signature of `transformDecoder` is modified to additionally provide the `newOffset` to the callback. We use both the original offset and the new offset in the error context.  

Fixes #755