Add a function to flatten transaction plan results (#1035)
#### Problem

When I've been looking at transaction plans, the `TransactionPlanResult` object is a bit too unweildy to be suitable for eg logging or quickly interpreting. 

![Screenshot 2025-11-21 at 14.53.47.png](https://app.graphite.com/user-attachments/assets/740285f4-ae3b-48d6-8119-2e4be7944497.png)

This is mostly because of its nesting and the fact it includes the transaction message and transaction objects.

#### Summary of Changes

This PR provides a function `flattenTransactionPlanResult` that flattens (depth first) a transaction plan result.

We can build a summary function on top of this.