Make Transaction optional in successful transaction plan result + add signature (#1043)
#### Problem

- The `transaction: Transaction` field being required means that you can't use for example a sending signer in a `TransactionExecutor`, as they only return a signature (mirroring wallet-standard sign+send)
- It is useful to have the signature of successful transactions in the `TransactionPlanResult`

#### Summary of Changes

This PR modifies the successful status to make `transaction` optional, and adds a new mandatory `Signature` field.

The function `successfulSingleTransactionPlanResult` now uses `getSignatureFromTransaction` to extract the signature from the transaction it receives, which it includes in the `SingleTransactionPlanResult`.

A new function `successfulSingleTransactionPlanResultFromSignature` can be used to create a successful `SingleTransactionPlanResult` with a signature but no transaction.