Accept any `Error` object in failed `SingleTransactionPlanResult` (#1056)
#### Problem

Currently, the failed status of `SinglePlanTransactionResults` accepts a `SolanaError` to capture the cause of the transaction failure.

Whilst internally to Kit we may have the ability to ensure `SolanaErrors` are thrown, this may not be the case for our consumers which have the implement the function that execute transactions and, therefore, find the "closest" `SolanaError` that matches their use-case.

#### Summary of Changes

This PR fixes this by loosening the error type we store in failed `SinglePlanTransactionResults` from `SolanaError` to any `Error` object.

This means we can continue to use the `isSolanaError(error, code)` helper to cherry-pick the use-cases we want to gracefully handle.