Allow signing a transaction that does not have TransactionWithLifetime (#1058)
#### Problem

I think we went a bit far here. It's important that `signTransaction` and `partiallySignTransaction` return the same `TTransaction extends Transaction` that is passed to them, so that they eg. preserve the lifetime. But I think it is ok to sign a `Transaction` that does not have a lifetime field attached to it. One place this is useful is a server that receives serialized transactions to sign, it should be able to sign the result of `transactionDecoder.decode(transactionBytes)` without concerning itself with that transaction's lifetime.

#### Summary of Changes

Remove `TransactionWithLifetime` from the required input type of `signTransaction` and `partiallySignTransaction`.