Throw on non-divisible transaction plan (#1155)
#### Problem

We currently offer a default implementation of `TransactionPlanExecutors` which requires a function to send a single transaction message. As such, this default implementation cannot handle `NonDivisibleSequentialTransactionPlans` since it would involve sending multiple transactions atomically — e.g. using Jito bundles.

The issue is the default implementation currently ignores the requirement of `NonDivisibleSequentialTransactionPlans` and sends them as if they were regular `SequentialTransactionPlans`. This is a mistake (made by me) as it invalid the atomicity of the `NonDivisibleSequentialTransactionPlans` and could lead to serious issues.

#### Summary of Changes

This PR fixes this issue by throwing a `SolanaError` when we encounter a `NonDivisibleSequentialTransactionPlan` in our default `TransactionPlanExecutor` implementation.

Importantly, it asserts that no `NonDivisibleSequentialTransactionPlan` exists in the potentially nested `TransactionPlan` provided _before_ trying to send any transaction. This avoid having a partially successful plan execution when we could have avoided it from the start.

#### Changeset reasoning

I went with a `patch` update here because, whilst this is a breaking change, it was not working as intended before. Therefore, I see this more as a bug fix than a breaking change but do let me know if you have another view on this.

#### Acknowledgements

Shout out to [Andy](https://linktr.ee/solandy) for flagging this in [one of his latest videos](https://www.youtube.com/watch?v=9S4ybrfiqUs).